# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger definition for MedRoom backend.",
        "title": "MedRoom API",
        "termsOfService": "https://swagger.io/terms/",
        "contact": {
            "name": "Martín Saavedra",
            "url": "https://www.swagger.io/support",
            "email": "martin.saavedra@mail.udp.cl"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/administradores-ti": {
            "get": {
                "description": "Lista todos los administradores-ti",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AdministradoresTi"
                ],
                "summary": "Lista de administradores-ti",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/SwaggerMessages.ListAdministradoresTiSwagger"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "description": "Genera un nuevo administrador_ti con los datos entregados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AdministradoresTi"
                ],
                "summary": "Agrega un nuevo administrador_ti",
                "parameters": [
                    {
                        "description": "AdministradorTi a agregar",
                        "name": "input_administrador_ti",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestMessages.AddNewAdministradorTiPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.AddNewAdministradorTiSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            }
        },
        "/administradores-ti/{uuid_administrador_ti}": {
            "get": {
                "description": "Obtiene un administrador_ti según su UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AdministradoresTi"
                ],
                "summary": "Obtiene un administrador_ti",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID del administrador_ti a buscar",
                        "name": "uuid_administrador_ti",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.GetOneAdministradorTiSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "description": "Modifica un administrador_ti con los datos entregados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AdministradoresTi"
                ],
                "summary": "Modifica un administrador_ti",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID del administrador_ti a modificar",
                        "name": "uuid_administrador_ti",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "AdministradorTi a modificar",
                        "name": "input_actualiza_administrador_ti",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestMessages.PutOneAdministradorTiPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.PutOneAdministradorTiSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un administrador_ti con los datos entregados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AdministradoresTi"
                ],
                "summary": "Elimina un administrador_ti",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID del administrador_ti a eliminar",
                        "name": "uuid_administrador_ti",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.DeleteAdministradorTiSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ApiHelpers.ResponseError": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "RequestMessages.AddNewAdministradorTiPayload": {
            "type": "object",
            "properties": {
                "apellidos_administrador_ti": {
                    "type": "string"
                },
                "correo_electronico_administrador_ti": {
                    "type": "string"
                },
                "hash_contrasena_administrador_ti": {
                    "type": "string"
                },
                "id_grupo": {
                    "type": "integer"
                },
                "id_rol": {
                    "type": "integer"
                },
                "nombres_administrador_ti": {
                    "type": "string"
                },
                "rut_administrador_ti": {
                    "type": "string"
                },
                "telefono_celular_administrador_ti": {
                    "type": "string"
                },
                "telefono_fijo_administrador_ti": {
                    "type": "string"
                }
            }
        },
...
        "RequestMessages.PutOneAdministradorTiPayload": {
            "type": "object",
            "properties": {
                "apellidos_administrador_ti": {
                    "type": "string"
                },
                "correo_electronico_administrador_ti": {
                    "type": "string"
                },
                "hash_contrasena_administrador_ti": {
                    "type": "string"
                },
                "id_grupo": {
                    "type": "integer"
                },
                "id_rol": {
                    "type": "integer"
                },
                "nombres_administrador_ti": {
                    "type": "string"
                },
                "rut_administrador_ti": {
                    "type": "string"
                },
                "telefono_celular_administrador_ti": {
                    "type": "string"
                },
                "telefono_fijo_administrador_ti": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.AddNewAdministradorTiResponse": {
            "type": "object",
            "properties": {
                "apellidos_administrador_ti": {
                    "type": "string"
                },
                "correo_electronico_administrador_ti": {
                    "type": "string"
                },
                "id_rol": {
                    "type": "integer"
                },
                "nombres_administrador_ti": {
                    "type": "string"
                },
                "rut_administrador_ti": {
                    "type": "string"
                },
                "telefono_celular_administrador_ti": {
                    "type": "string"
                },
                "telefono_fijo_administrador_ti": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.DeleteAdministradorTiResponse": {
            "type": "object",
            "properties": {
                "apellidos_administrador_ti": {
                    "type": "string"
                },
                "correo_electronico_administrador_ti": {
                    "type": "string"
                },
                "id_rol": {
                    "type": "integer"
                },
                "nombres_administrador_ti": {
                    "type": "string"
                },
                "rut_administrador_ti": {
                    "type": "string"
                },
                "telefono_celular_administrador_ti": {
                    "type": "string"
                },
                "telefono_fijo_administrador_ti": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.GetOneAdministradorTiResponse": {
            "type": "object",
            "properties": {
                "apellidos_administrador_ti": {
                    "type": "string"
                },
                "correo_electronico_administrador_ti": {
                    "type": "string"
                },
                "nombres_administrador_ti": {
                    "type": "string"
                },
                "rol_administrador_ti": {
                    "$ref": "#/definitions/ResponseMessages.GetOneRolResponse"
                },
                "rut_administrador_ti": {
                    "type": "string"
                },
                "telefono_celular_administrador_ti": {
                    "type": "string"
                },
                "telefono_fijo_administrador_ti": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.GetOneRolResponse": {
            "type": "object",
            "properties": {
                "nombre_rol": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.ListAdministradoresTiResponse": {
            "type": "object",
            "properties": {
                "apellidos_administrador_ti": {
                    "type": "string"
                },
                "correo_electronico_administrador_ti": {
                    "type": "string"
                },
                "nombres_administrador_ti": {
                    "type": "string"
                },
                "rol_administrador_ti": {
                    "$ref": "#/definitions/ResponseMessages.GetOneRolResponse"
                },
                "rut_administrador_ti": {
                    "type": "string"
                },
                "telefono_celular_administrador_ti": {
                    "type": "string"
                },
                "telefono_fijo_administrador_ti": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.PutOneAdministradorTiResponse": {
            "type": "object",
            "properties": {
                "apellidos_administrador_ti": {
                    "type": "string"
                },
                "correo_electronico_administrador_ti": {
                    "type": "string"
                },
                "hash_contrasena_administrador_ti": {
                    "type": "string"
                },
                "id_rol": {
                    "type": "integer"
                },
                "nombres_administrador_ti": {
                    "type": "string"
                },
                "rut_administrador_ti": {
                    "type": "string"
                },
                "telefono_celular_administrador_ti": {
                    "type": "string"
                },
                "telefono_fijo_administrador_ti": {
                    "type": "string"
                }
            }
        },
...
        "SwaggerMessages.AddNewAdministradorTiSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.AddNewAdministradorTiResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.DeleteAdministradorTiSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.DeleteAdministradorTiResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.GetOneAdministradorTiSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.GetOneAdministradorTiResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.ListAdministradoresTiSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.ListAdministradoresTiResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.PutOneAdministradorTiSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.PutOneAdministradorTiResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    }
}